Create a hybrid_layers module for neuron-level control of layers
